+++ src/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/[lessonSlug]/downloads/success/Components/DownloadSuccessView.tsx
@@ -8,6 +8,7 @@ import {
   OakIcon,
   OakSpan,
 } from "@oaknational/oak-components";
+import { useEffect } from "react";
 
 import { LessonList } from "@/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/Components/LessonList";
 import { DownloadSuccessHeader } from "@/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/[lessonSlug]/Components/DownloadSuccessHeader/DownloadSuccessHeader";
@@ -18,6 +19,7 @@ import { resolveOakHref } from "@/common-lib/urls";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import type { LessonListSchema } from "@/node-lib/curriculum-api-2023/shared.schema";
 import { KeyStageTitleValueType } from "@/browser-lib/avo/Avo";
+import { useOakNotificationsContext } from "@/context/OakNotifications/useOakNotificationsContext";
 import { getUnitDownloadFileId } from "@/utils/getUnitDownloadFileId";
 
 type DownloadSuccessViewLesson = {
@@ -61,6 +63,8 @@ export function DownloadSuccessView({
   const { track } = useAnalytics();
   const { onwardContentSelected, unitDownloadInitiated } = track;
 
+  const { setCurrentToastProps } = useOakNotificationsContext();
+
   const {
     setShowDownloadMessage,
     setDownloadError,
@@ -72,6 +76,17 @@ export function DownloadSuccessView({
   const isGeorestrictedUnit = lessons.some(
     (l) => "geoRestricted" in l && l.geoRestricted,
   );
+
+  useEffect(() => {
+    setCurrentToastProps({
+      message: "Download started. This may take a few minutes",
+      variant: "success",
+      autoDismiss: true,
+      showClose: true,
+      showIcon: true,
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <>