+++ src/components/TeacherViews/LessonDownloads.view.tsx
@@ -55,6 +55,7 @@ import { LegacyCopyrightContent } from "@/node-lib/curriculum-api-2023/shared.sc
 import { LessonDownloadRegionBlocked } from "@/components/TeacherComponents/LessonDownloadRegionBlocked/LessonDownloadRegionBlocked";
 import { resolveOakHref } from "@/common-lib/urls";
 import { useComplexCopyright } from "@/hooks/useComplexCopyright";
+import { useOakNotificationsContext } from "@/context/OakNotifications/useOakNotificationsContext";
 
 type BaseLessonDownload = {
   expired: boolean | null;
@@ -98,6 +99,7 @@ type LessonDownloadsProps =
 
 export function LessonDownloads(props: LessonDownloadsProps) {
   const { lesson } = props;
+  const { setCurrentToastProps } = useOakNotificationsContext();
   const router = useRouter();
   const {
     lessonTitle,
@@ -266,7 +268,7 @@ export function LessonDownloads(props: LessonDownloadsProps) {
         schoolName,
         schoolOption,
         onwardContent,
-        emailSupplied: data?.email ? true : false,
+        emailSupplied: !!data?.email,
         platform: "owa",
         product: "teacher lesson resources",
         engagementIntent: "use",
@@ -288,6 +290,13 @@ export function LessonDownloads(props: LessonDownloadsProps) {
       setApiError(
         "There was an error downloading your files. Please try again.",
       );
+      setCurrentToastProps({
+        message:
+          "Something went wrong with the download. Try refreshing the page.",
+        variant: "error",
+        autoDismiss: false,
+        showIcon: true,
+      });
     }
   };
 
@@ -345,7 +354,7 @@ export function LessonDownloads(props: LessonDownloadsProps) {
             $mb={"spacing-24"}
           />
         </OakBox>
-        {showGeoBlocked ? (
+        {showGeoBlocked && (
           <LessonDownloadRegionBlocked
             lessonName={lessonTitle}
             lessonSlug={lessonSlug}
@@ -358,7 +367,8 @@ export function LessonDownloads(props: LessonDownloadsProps) {
               unitSlug: unitSlug!,
             })}
           />
-        ) : isDownloadSuccessful ? (
+        )}
+        {!showGeoBlocked && isDownloadSuccessful && (
           <DownloadConfirmation
             lessonSlug={lessonSlug}
             lessonTitle={lessonTitle}
@@ -389,7 +399,8 @@ export function LessonDownloads(props: LessonDownloadsProps) {
             isLegacy={isLegacyDownload}
             lessonReleaseDate={lessonReleaseDate ?? "unreleased"}
           />
-        ) : (
+        )}
+        {!showGeoBlocked && !isDownloadSuccessful && (
           <DownloadPageWithAccordion
             loginRequired={loginRequired ?? false}
             geoRestricted={geoRestricted ?? false}
@@ -431,9 +442,7 @@ export function LessonDownloads(props: LessonDownloadsProps) {
             cta={
               <OakPrimaryButton
                 type="button"
-                onClick={
-                  (event) => void form.handleSubmit(onFormSubmit)(event) // https://github.com/orgs/react-hook-form/discussions/8622}
-                }
+                onClick={(event) => void form.handleSubmit(onFormSubmit)(event)} // https://github.com/orgs/react-hook-form/discussions/8622}
                 iconName={"download"}
                 isLoading={
                   isAttemptingDownload || !hubspotLoaded // show loading state when waiting for latest school values to be populated from hubspot