+++ src/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/downloads.tsx
@@ -86,12 +86,11 @@ export const getStaticProps: GetStaticProps<
 
       let curriculumData;
       try {
-        curriculumData =
-          await curriculumApi2023.lessonDownloads<LessonDownloadsPageData>({
-            programmeSlug,
-            unitSlug,
-            lessonSlug,
-          });
+        curriculumData = await curriculumApi2023.lessonDownloads({
+          programmeSlug,
+          unitSlug,
+          lessonSlug,
+        });
       } catch (innerError) {
         allowNotFoundError(innerError);
       }